sendgrid/python/requirements.txt
sendgrid
google-cloud-secret-manager